requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Click==7.0
decorator==4.4.0
Flask==1.1.1
flask-jsontools==0.1.5
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
importlib-metadata==0.23
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==7.2.0
passlib==1.7.1
pyasn1==0.4.7
pysmb==1.1.27
python-dotenv==0.10.3
pytz==2019.3
redis==3.3.11
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-Utc==0.10.0
SQLAlchemy-Utils==0.34.2
validators==0.14.0
vine==1.3.0
Werkzeug==0.16.0
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
WTForms-SQLAlchemy==0.1
zipp==0.6.0
